+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 + Allow `React.memo` to specify a custom compare function [PR #260](https://github.com/shadaj/slinky/pull/260)
 
 ### Bug Fixes
++ Allow `component` defined in a functional component to have any access control [PR #262](https://github.com/shadaj/slinky/pull/262)
 + Fix incorrect typing of `SyntheticMouseEvent.clientX` [PR #261](https://github.com/shadaj/slinky/pull/261)
 
 ## [v0.6.0](https://slinky.dev)